RavenDB-11373 Replay Tx logic moved to helper
